+++ src/main/java/com/singsongchanson/domain/asset/command/application/service/AssetCommandService.java
@@ -3,8 +3,8 @@ package com.singsongchanson.domain.asset.command.application.service;
 import com.amazonaws.services.s3.AmazonS3Client;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.singsongchanson.domain.asset.command.application.dto.AssetResponseDTO;
-import com.singsongchanson.domain.asset.domain.entity.Asset;
-import com.singsongchanson.domain.asset.domain.repository.AssetRepository;
+import com.singsongchanson.domain.asset.command.domain.entity.Asset;
+import com.singsongchanson.domain.asset.command.domain.repository.AssetRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
