+++ src/main/java/com/singsongchanson/domain/asset/command/domain/repository/AssetRepository.java
@@ -1,6 +1,6 @@
-package com.singsongchanson.domain.asset.domain.repository;
+package com.singsongchanson.domain.asset.command.domain.repository;
 
-import com.singsongchanson.domain.asset.domain.entity.Asset;
+import com.singsongchanson.domain.asset.command.domain.entity.Asset;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
