+++ src/main/java/com/singsongchanson/domain/user/query/application/service/UserQueryService.java
@@ -3,7 +3,7 @@ package com.singsongchanson.domain.user.query.application.service;
 import com.singsongchanson.domain.user.command.application.dto.FindUserResponseDTO;
 import com.singsongchanson.domain.user.command.domain.aggregate.entity.enumtype.SocialType;
 import com.singsongchanson.domain.user.query.infrastructure.repository.UserQueryRepository;
-import com.singsongchanson.global.security.jwt.domain.repository.TokenRepository;
+import com.singsongchanson.global.security.jwt.command.domain.repository.TokenRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -12,7 +12,6 @@ import org.springframework.stereotype.Service;
 public class UserQueryService {
 
     private final UserQueryRepository userQueryRepository;
-    private final TokenRepository tokenRepository;
 
     public FindUserResponseDTO findUserByUserNo(Long userNo) {
