+++ src/main/java/com/singsongchanson/global/security/jwt/command/application/service/JwtProvider.java
@@ -3,10 +3,10 @@ package com.singsongchanson.global.security.jwt.command.application.service;
 import com.singsongchanson.domain.user.command.application.dto.FindUserResponseDTO;
 import com.singsongchanson.global.configuration.AppProperties;
 import com.singsongchanson.global.security.UserPrincipal;
-import com.singsongchanson.global.security.jwt.domain.aggregate.entity.Token;
-import com.singsongchanson.global.security.jwt.domain.aggregate.vo.UserVO;
-import com.singsongchanson.global.security.jwt.domain.repository.TokenRepository;
-import com.singsongchanson.global.security.jwt.domain.service.RequestUserDomainService;
+import com.singsongchanson.global.security.jwt.command.domain.aggregate.entity.Token;
+import com.singsongchanson.global.security.jwt.command.domain.aggregate.vo.UserVO;
+import com.singsongchanson.global.security.jwt.command.domain.repository.TokenRepository;
+import com.singsongchanson.global.security.jwt.command.domain.service.RequestUserDomainService;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
