+++ src/main/java/com/singsongchanson/global/security/jwt/command/infrastructure/service/RequestUserInfraService.java
@@ -2,7 +2,7 @@ package com.singsongchanson.global.security.jwt.command.infrastructure.service;
 
 import com.singsongchanson.domain.user.command.application.dto.FindUserResponseDTO;
 import com.singsongchanson.domain.user.query.application.service.UserQueryService;
-import com.singsongchanson.global.security.jwt.domain.service.RequestUserDomainService;
+import com.singsongchanson.global.security.jwt.command.domain.service.RequestUserDomainService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
